requirements.txt
django
djangorestframework
markdown
django-filter
braintree
psycopg2==2.5.4
awsebcli
